environments/photon_scene: trace photons

+++ src/core/sampler.rs
@@ -6,7 +6,7 @@ pub struct Point2D {
 }
 
 pub trait Sampler {
-    fn samples(num_samples: u8) -> Vec<Point2D>;
+    fn samples(num_samples: u16) -> Vec<Point2D>;
 
     fn hemisphere_sampler(&self, e: f32) -> Vec<Vector>;
 }
